### 2019-01-15 #1174 @yochannah
@andrewthomasjones No worries - thankfully our reviewer volunteer list does have a lot of Machine learning + R users so hopefully it won't be too hard to find someone (famous last words, I know! 😆 )

@ja-thomas @schalkdaniel - can we interest you in reviewing this paper by any chance?
