### 2019-01-17 #1174 @ja-thomas
happy to do it. I guess @schalkdaniel might be suited a bit better, due to his RCPP experience, but if he declines I can do it as well :)
