### 2019-01-21 #1193 @whedon
[REVIEW]: BoltzMM: an R package for maximum pseudolikelihood estimation of fully-visible   Boltzmann machines
**Submitting author:** @andrewthomasjones (<a href="http://orcid.org/0000-0003-1324-703X">Andrew Jones</a>)
**Repository:** <a href="https://github.com/andrewthomasjones/BoltzMM" target ="_blank">https://github.com/andrewthomasjones/BoltzMM</a>
**Version:** 0.1.3
