### 2019-01-22 48547c5
added sticker
