+++ README.md
@@ -4,6 +4,7 @@
 mirror](http://cranlogs.r-pkg.org/badges/BoltzMM)](https://CRAN.R-project.org/package=BoltzMM)
 [![Build
 Status](https://travis-ci.org/andrewthomasjones/BoltzMM.svg?branch=master)](https://travis-ci.org/andrewthomasjones/BoltzMM)
+[![status](http://joss.theoj.org/papers/23eb189a5e0bdd2b51f668621abcc75a/status.svg)](http://joss.theoj.org/papers/23eb189a5e0bdd2b51f668621abcc75a)
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
@@ -35,9 +36,9 @@ install.packages("BoltzMM", repos='http://cran.us.r-project.org')
 ```
 
 An archival build of `BoltzMM` is available at
-<https://zenodo.org/record/1339352>. Manual installation instructions
-can be found within the *R* installation and administration manual
-<https://cran.r-project.org/doc/manuals/r-release/R-admin.html>.
+<http://doi.org/10.5281/zenodo.2538256>. Manual installation
+instructions can be found within the *R* installation and administration
+manual <https://cran.r-project.org/doc/manuals/r-release/R-admin.html>.
 
 ## Examples
 
@@ -170,12 +171,12 @@ test_results
 #> 
 #> $Mmat_z
 #>            [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
-#> [1,]        NaN -0.6596632 -1.3152156 -2.0181850  0.1936413  3.8587216
-#> [2,] -0.6596632        NaN -0.9536614 -2.0795939 -0.4947831  6.1625534
-#> [3,] -1.3152156 -0.9536614        NaN  1.3258059 -1.0861345  2.6332797
-#> [4,] -2.0181850 -2.0795939  1.3258059        NaN  3.0115924  0.4798541
-#> [5,]  0.1936413 -0.4947831 -1.0861345  3.0115924        NaN -1.2899547
-#> [6,]  3.8587216  6.1625534  2.6332797  0.4798541 -1.2899547        NaN
+#> [1,]         NA -0.6596632 -1.3152156 -2.0181850  0.1936413  3.8587216
+#> [2,] -0.6596632         NA -0.9536614 -2.0795939 -0.4947831  6.1625534
+#> [3,] -1.3152156 -0.9536614         NA  1.3258059 -1.0861345  2.6332797
+#> [4,] -2.0181850 -2.0795939  1.3258059         NA  3.0115924  0.4798541
+#> [5,]  0.1936413 -0.4947831 -1.0861345  3.0115924         NA -1.2899547
+#> [6,]  3.8587216  6.1625534  2.6332797  0.4798541 -1.2899547         NA
 #> [7,]  0.5671620  0.5877623  5.8430378 -0.9769979  1.5757127 -1.0129338
 #> [8,]  0.3126387 -3.4715041 -0.9287578  4.0101249  0.7587521  2.3224311
 #>            [,7]       [,8]
@@ -185,16 +186,16 @@ test_results
 #> [4,] -0.9769979  4.0101249
 #> [5,]  1.5757127  0.7587521
 #> [6,] -1.0129338  2.3224311
-#> [7,]        NaN  2.2571849
-#> [8,]  2.2571849        NaN
+#> [7,]         NA  2.2571849
+#> [8,]  2.2571849         NA
 #> 
 #> $Mmat_p
 #>              [,1]         [,2]         [,3]         [,4]        [,5]
-#> [1,]          NaN 5.094700e-01 1.884374e-01 4.357200e-02 0.846456748
[…]
