### 2019-01-29 andrewthomasjones/BoltzMM#3 @schalkdaniel
JOSS Review: Testing the package after first install fails
After installing the package and running the tests as documented in the Readme I get errors (see below). Running the test a second and third times looks ok.

Anyway, the output in the error messages looks correct except for rounding errors. So this may depend on the local print settings?

```r
> devtools::install_github('andrewthomasjones/BoltzMM')
> testthat::test_package("BoltzMM")
#> ── 1. Failure: Test fitfvbm  (@test-cpp.R#14)  ─────────────────────────────────
#> `fitfvbmResult` has changed from known value recorded in './ffitfvbm'.
#> 6/15 mismatches
#> x[2]: "[1] -19.2"
#> y[2]: "[1] -19.20066"
#>
#> x[5]: "[1]  0.2473 -0.1063 -0.4735"
#> y[5]: "[1]  0.2472515 -0.1063203 -0.4735163"
#>
#> x[8]: "        [,1]     [,2]    [,3]"
#> y[8]: "           [,1]        [,2]       [,3]"
#>
#> x[9]: "[1,] 0.00000  0.02096  0.1064"
#> y[9]: "[1,] 0.00000000  0.02095682  0.1063595"
#>
#> x[10]: "[2,] 0.02096  0.00000 -0.2473"
#> y[10]: "[2,] 0.02095682  0.00000000 -0.2472607"
#>
#> ── 2. Failure: Test fvbmpartiald (@test-cpp.R#48)  ─────────────────────────────
#> `fvbmpartialdResult` has changed from known value recorded in './fvbmpartiald'.
#> 5/9 mismatches
#> x[2]: "[1]  -2.421 -35.192 -48.432"
#> y[2]: "[1]  -2.420997 -35.192266 -48.432288"
#>
#> x[5]: "      [,1]   [,2]   [,3]"
#> y[5]: "          [,1]      [,2]      [,3]"
#>
#> x[6]: "[1,]   0.0 -24.90  14.60"
#> y[6]: "[1,]   0.00000 -24.90402  14.59706"
#>
#> x[7]: "[2,] -24.9   0.00 -17.06"
#> y[7]: "[2,] -24.90402   0.00000 -17.05561"
#>
#> x[8]: "[3,]  14.6 -17.06   0.00"
#> y[8]: "[3,]  14.59706 -17.05561   0.00000"
#>
#> ── 3. Failure: Test fvbmcov (@test-cpp.R#80)  ──────────────────────────────────
#> `fvbmcovResult` has changed from known value recorded in './fvbmcov'.
#> 7/7 mismatches
#> x[1]: "         [,1]    [,2]     [,3]      [,4]     [,5]      [,6]"
#> y[1]: "            [,1]        [,2]        [,3]         [,4]        [,5]
#> y[1]:  [,6]"
#>
#> x[2]: "[1,]  1.02193 -0.3985  0.01122 -0.323185  0.02596 -0.125816"
#> y[2]: "[1,]  1.02192628 -0.39848488  0.01122311 -0.323185292  0.02596410 -0.1258
#> y[2]: 15718"
#>
#> x[3]: "[2,] -0.39848  1.9154 -0.08760  0.393365 -0.15700  0.327725"
#> y[3]: "[2,] -0.39848488  1.91544879 -0.08760205  0.393364949 -0.15700126  0.3277
#> y[3]: 25222"
#>
#> x[4]: "[3,]  0.01122 -0.0876  1.39409 -0.053356  0.14167 -0.158487"
#> y[4]: "[3,]  0.01122311 -0.08760205  1.39408988 -0.053355837  0.14167054 -0.1584
#> y[4]: 87282"
#>
#> x[5]: "[4,] -0.32319  0.3934 -0.05336  1.337065 -0.02986  0.006558"
#> y[5]: "[4,] -0.32318529  0.39336495 -0.05335584  1.337064987 -0.02986267  0.0065
#> y[5]: 58303"
#>
#> ── 4. Failure: Check that fvbmHess calculates Hesssian Correctly. (@test-rfunct.
#> `HessResult` has changed from known value recorded in './fvbmHess'.
#> 7/7 mismatches
#> x[1]: "        [,1]   [,2]    [,3]     [,4]     [,5]     [,6]"
#> y[1]: "           [,1]      [,2]       [,3]        [,4]        [,5]        [,6]"
#>
#> x[2]: "[1,] -98.130   0.00   0.000   -8.312   23.688    0.000"
#> y[2]: "[1,] -98.130063   0.00000   0.000000   -8.311656   23.688344    0.000000"
#>
#> x[3]: "[2,]   0.000 -79.94   0.000   11.116    0.000   26.845"
#> y[3]: "[2,]   0.000000 -79.94001   0.000000   11.116182    0.000000   26.845137"
#>
#> x[4]: "[3,]   0.000   0.00 -92.373    0.000    8.033   -3.248"
#> y[4]: "[3,]   0.000000   0.00000 -92.373399    0.000000    8.032654   -3.247524"
#>
#> x[5]: "[4,]  -8.312  11.12   0.000 -178.070    5.870  -13.878"
#> y[5]: "[4,]  -8.311656  11.11618   0.000000 -178.070070    5.869937  -13.878397"
#>
#> ── 5. Failure: Check that marginpfvbm calculates marginal probailities Correctly
#> `marginResult` has changed from known value recorded in './marginpfvbm'.
#> 1/1 mismatches
#> x[1]: "[1] 0.5382 0.7415 0.6460"
#> y[1]: "[1] 0.5382340 0.7414746 0.6459500"
#>
#> ── 6. Failure: Check that fvbmtests calculates scores and p-values correctly. (@
#> `testResult` has changed from known value recorded in './fvbmtests'.
#> 10/18 mismatches
#> x[2]: "[1]  0.7068 12.6493  7.5033"
#> y[2]: "[1]  0.7068241 12.6492636  7.5033160"
#>
#> x[5]: "[1] 4.797e-01 1.129e-36 6.222e-14"
#> y[5]: "[1] 4.796758e-01 1.129283e-36 6.222331e-14"
#>
#> x[8]: "      [,1]   [,2]   [,3]"
#> y[8]: "        [,1]      [,2]      [,3]"
#>
#> x[9]: "[1,]    NA 3.2899 4.3007"
#> y[9]: "[1,]      NA 3.2899196 4.3007296"
#>
#> x[10]: "[2,] 3.290     NA 0.9492"
#> y[10]: "[2,] 3.28992        NA 0.9492276"
#>
#> ══ testthat results  ═══════════════════════════════════════════════════════════
#> OK: 5 SKIPPED: 0 FAILED: 6
#> 1. Failure: Test fitfvbm  (@test-cpp.R#14)
#> 2. Failure: Test fvbmpartiald (@test-cpp.R#48)
#> 3. Failure: Test fvbmcov (@test-cpp.R#80)
#> 4. Failure: Check that fvbmHess calculates Hesssian Correctly. (@test-rfunct.R#17)
#> 5. Failure: Check that marginpfvbm calculates marginal probailities Correctly. (@test-rfunct.R#42)
#> 6. Failure: Check that fvbmtests calculates scores and p-values correctly. (@test-rfunct.R#57)
#>
#> Error: testthat unit tests failed
```
