### 2019-01-29 andrewthomasjones/BoltzMM#3 @ja-thomas
Same problems here, after cloning the repository, and running the test locally.

Additionally the instructions given in the README do not work for me stating `Error: No tests found for BoltzMM`. I need to clone `document()` and `test()` manually.
