### 2019-01-29 #1193 @ja-thomas
Major comments:

- The second author of the paper "Jessica J. Bagnall" does not seem to have commits to the repository. It might be a good idea to specify the contributions of every author in the README.

- The tests cannot be reproduced with the instructions in the README

- 5 out of 11 tests fail when running locally, after calling `devtools::document()` and `devtools::test()`, all with `XXX has changed from known value recorded in XXX`. (See https://github.com/andrewthomasjones/BoltzMM/issues/3)

Minor comments:

- Using the shortcut `T` for `TRUE` is rather bad practice in R since it can be easily overwritten as a variable. (As done in the installation instructions)
- `knn.impute` produces warnings in the example
- Installation works (with Ubuntu 18.04). Since the version is on CRAN, installation on windows should have been checked by the WINbuilder. It still might be a good idea to add continuous integration for Linux (with Travis) and Windows (with Appveyor). Especially considering the above problems with reproducing the tests
- A reference to the paper given the theoretical framework for this method would be nice. (Or the JOSS paper, since it contains/references the required backround).
