+++ README.Rmd
@@ -12,10 +12,10 @@ knitr::opts_chunk$set(
 
 
 
-<img src="http://www.r-pkg.org/badges/version-last-release/BoltzMM"></img></a>
-[![Downloads from the RStudio CRAN mirror](http://cranlogs.r-pkg.org/badges/BoltzMM)](https://CRAN.R-project.org/package=BoltzMM)
-[![Build Status](https://travis-ci.org/andrewthomasjones/BoltzMM.svg?branch=master)](https://travis-ci.org/andrewthomasjones/BoltzMM)
-[![status](http://joss.theoj.org/papers/23eb189a5e0bdd2b51f668621abcc75a/status.svg)](http://joss.theoj.org/papers/23eb189a5e0bdd2b51f668621abcc75a)
+<!-- <img src="http://www.r-pkg.org/badges/version-last-release/BoltzMM"></img></a> -->
+<!-- [![Downloads from the RStudio CRAN mirror](http://cranlogs.r-pkg.org/badges/BoltzMM)](https://CRAN.R-project.org/package=BoltzMM) -->
+<!-- [![Build Status](https://travis-ci.org/andrewthomasjones/BoltzMM.svg?branch=master)](https://travis-ci.org/andrewthomasjones/BoltzMM) -->
+<!-- [![status](http://joss.theoj.org/papers/23eb189a5e0bdd2b51f668621abcc75a/status.svg)](http://joss.theoj.org/papers/23eb189a5e0bdd2b51f668621abcc75a) -->
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
@@ -95,18 +95,20 @@ Example with real data from https://hal.archives-ouvertes.fr/hal-01927188v1.
 library(bnstruct)
 library(BoltzMM)
 
-#load data
+# Load data
 data(senate)
 
 # Turn data into a matrix
 senate_data <- as.matrix(senate)
 
-#recode Yes as 1, and No as -1
-senate_data[senate_data=="Yes"]<-1
-senate_data[senate_data=="No"]<--1
+# Recode Yes as 1, and No as -1, and make the "NA"s proper NA values
+senate_data[senate=="Yes"] <- 1
+senate_data[senate=="No"] <- -1
 
-# Conduct imputation for missing data
-imp_data <- knn.impute(senate_data,k=1)
+# Conduct imputation
+imp_data <- knn.impute(suppressWarnings(matrix(as.numeric(senate_data),
+                                        dim(senate_data))),
+                       k=1)
 
 # No governement - using as reference level
 data_nogov <- imp_data[,-1]
@@ -129,7 +131,20 @@ test_results<-fvbmtests(data_nogov,model,nullmodel)
 test_results
 ```
 
-For more examples see individual help files.
+For more examples, see individual help files.
+
+## Reference to package
+
+If you find this package useful in your work, then please follow the usual `R` instructions for citing the package in your publications. That is, follow the instructions from `citation('BoltzMM')`.
+
+```{r citation}
+# Citation instructions
+citation('BoltzMM')
+```
+
+## Contributions
+
+The `BoltzMM` package is authored by Andrew T. Jones [@andrewthomasjones](https://github.com/andrewthomasjones), Hien D. Nguyen (@hiendn), and Jessica J. Bagnall. 
 
 ## Unit testing
 
@@ -147,16 +162,6 @@ library(testthat)
 test_package("BoltzMM")
 ```
 
-## Reference to package
-
-If you find this package useful in your work, then please follow the usual `R` instructions for citing the package in your publications. That is, follow the instructions from `citation('BoltzMM')`.
-
-```{r citation}
-# Citation instructions
-citation('BoltzMM')
-```
-
 ## Bug reporting and contributions
 
[…]
