### 2019-01-30 24c003d
Update author statements ++
