### 2019-01-30 #1193 @hiendn
Hi @ja-thomas, thank you for taking the time to review our paper! We very much appreciate your time and effort! 😊

I will reply to some of your comments now, and @andrewthomasjones will reply to the others at a later stage.

> The second author of the paper "Jessica J. Bagnall" does not seem to have commits to the repository. It might be a good idea to specify the contributions of every author in the README.

Unfortunately, Jessica is not actively on GitHub (we will persuade her, yet!) and has therefore made most of her contributions via commits by @andrewthomasjones and I. She has made numerous contributions to the paper via the documentation of the package, the `senate` data set that she had supplied and analysed from her paper https://hal.archives-ouvertes.fr/hal-01927188v1. As well as with identifying and helping to fix various bugs that were in previous versions of the code.

We have now included an **Authorship statement** section in the README in order to highlight the contributions of each of the three co-authors of the package.

> knn.impute produces warnings in the example

Unfortunately, this cannot be help. No matter how the data is processed or preprocessed before conducting imputation, `NA`s will be created since there is presence of missingness. Thus, a warning **MUST** be signalled at some part in time in the workflow. Since it is a little bit unsightly to have *warning* messages in the README, we have opted to use a `suppressWarnings()` command in order to make the output more pleasing to the eye.

> A reference to the paper given the theoretical framework for this method would be nice. (Or the JOSS paper, since it contains/references the required backround).

This was a great idea! We have now included a **Technical references** section in the README to address this comment.

> There is no citation information in the README. At least the same information as `citation("BoltzMM")` should be given in the README

Again, great idea! We have now included a **Reference to package** section in the README that contains exactly the output of the command `citation("BoltzMM")` as you have suggested.

Thanks again for your time, @ja-thomas. It is very much appreciated.
