### 2019-01-30 #1193 @ja-thomas
great, thanks for the quick reply and fix.

I guess the only point open from my side are the issues with the tests discussed here https://github.com/andrewthomasjones/BoltzMM/issues/3. When this is resolved everything is fine from my side.
