### 2019-01-31 7c1633d
remove superflous ad incorrect testing instructions
