### 2019-01-31 58ed583
Revert "remove superflous ad incorrect testing instructions"
This reverts commit 7c1633de32f201d2555d43608ded448d68f96cdc.
