+++ README.Rmd
@@ -1,0 +1,176 @@
+---
+output: github_document
+---
+
+```{r, echo = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "README-"
+)
+```
+
+
+
+<img src="http://www.r-pkg.org/badges/version-last-release/BoltzMM"></img></a>
+[![Downloads from the RStudio CRAN mirror](http://cranlogs.r-pkg.org/badges/BoltzMM)](https://CRAN.R-project.org/package=BoltzMM)
+[![Build Status](https://travis-ci.org/andrewthomasjones/BoltzMM.svg?branch=master)](https://travis-ci.org/andrewthomasjones/BoltzMM)
+[![status](http://joss.theoj.org/papers/23eb189a5e0bdd2b51f668621abcc75a/status.svg)](http://joss.theoj.org/papers/23eb189a5e0bdd2b51f668621abcc75a)
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+```{r, echo = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "README-"
+)
+```
+
+# BoltzMM
+
+The BoltzMM package allows for computation of probability mass functions of fully-visible Boltzmann machines (FVBMs) via `pfvbm` and `allpfvbm`.
+Random data can be generated using `rfvbm`. Maximum pseudolikelihood estimation of parameters via the MM algorithm can be conducted using `fitfvbm`.
+Computation of partial derivatives and Hessians can be performed via `fvbmpartiald` and `fvbmHessian`.
+Covariance estimation and normal standard errors can be computed using `fvbmcov` and `fvbmstderr`.
+
+## Installation
+
+If `devtools` has already been installed, then the most current build of `BoltzMM` can be obtained via the command:
+
+```{r GH-install, eval=F}
+devtools::install_github('andrewthomasjones/BoltzMM',build_vignettes = T)
+```
+
+The latest stable build of `BoltzMM` can be obtain from CRAN via the command:
+
+```{r CRAN-install, eval = F}
+install.packages("BoltzMM", repos='http://cran.us.r-project.org')
+```
+
+An archival build of `BoltzMM` is available at http://doi.org/10.5281/zenodo.2538256. Manual installation instructions can be found within the *R* installation and administration manual https://cran.r-project.org/doc/manuals/r-release/R-admin.html.
+
+## Examples
+
+Compute the probability of every length n=3 binary spin vector under bvec and Mmat:
+
+```{r example, eval = T}
+library(BoltzMM)
+set.seed(1)
+
+bvec <- c(0,0.5,0.25)
+Mmat <- matrix(0.1,3,3) - diag(0.1,3,3)
+allpfvbm(bvec,Mmat)
+```
+
+Generate num=1000 random strings of n=3 binary spin variables under bvec and Mmat.
+```{r example2, eval = T}
+library(BoltzMM)
+set.seed(1)
+
+num <- 1000
+bvec <- c(0,0.5,0.25)
+Mmat <- matrix(0.1,3,3) - diag(0.1,3,3)
+data <- rfvbm(num,bvec,Mmat)
+
+head(data)
+```
+
+Fit a fully visible Boltzmann machine to data, starting from parameters bvec and Mmat.
+ 
+```{r example3, eval = T}
+library(BoltzMM)
+set.seed(1)
+
+bvec <- c(0,0.5,0.25)
+Mmat <- matrix(0.1,3,3) - diag(0.1,3,3)
+data <- rfvbm(num,bvec,Mmat)
+
+fitfvbm(data,bvec,Mmat)
+```
+
[…]
