### 2019-01-31 4c3b8f5
removed superflous and incorrect test instructions
