### 2019-01-31 andrewthomasjones/BoltzMM#3 @andrewthomasjones
Hi Guys,

Thanks for reviewing our software.

-----------
This was a very interesting bug. It didn't occur on CRAN build or through TravisCI, but it is nevertheless clearly relatively common.

I was unable to replicate the bug directly but as far as I can tell there is something system/OS/compiler/? specific about the conversion of C++ data into R that leads to these apparent rounding errors. The simplest and most robust solution I could think of was turning down the sensitivity of equality tests so they don't fail in these cases. I don't think this loss of precision will compromise their function as unit tests.

I hope this solution works for you, as I was unable to directly replicate the fault I'm not 100% sure this fix will definitely work [fingers crossed though :) ].

----------
I removed the incorrect testing instructions from the readme file as they are largely superfluous.

Thanks again,

Andrew
