### 2019-01-31 e7ec11a
Replace T and F by TRUE and FALSE
