### 2019-01-31 5ba8def
Replace T and F by TRUE and FALSE
