### 2019-02-06 andrewthomasjones/BoltzMM#3 @ja-thomas
The testing setup seems kind of weird.

The reason it works after calling test() the second time is that the objects you test against are changed/overwritting when calling the tests

```
Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git checkout -- <file>..." to discard changes in working directory)

	modified:   tests/testthat/ffitfvbm
	modified:   tests/testthat/fvbmHess
	modified:   tests/testthat/fvbmcov
	modified:   tests/testthat/fvbmpartiald
	modified:   tests/testthat/fvbmtests
	modified:   tests/testthat/marginpfvbm
```

is that the intendet behaviour?

Same issues with the tests unfortunately.

Generally I'm not sure if you actually want to unit test the correctness by checking for identical objects, but rather check that the structure of the result is correct, correct dimensions etc. and that edge cases work without creating errors, or throw (intendet) error messages
