### 2019-02-08 3998332
test files should no longer overwrite in folder when cloned
