### 2019-02-08 53123f7
additional, more fundamental tests of output
