### 2019-02-08 da3331b
remove stored value tests
