### 2019-02-08 andrewthomasjones/BoltzMM#3 @andrewthomasjones
That is unexpected behaviour.

Thanks for the advice on unit testing. I've modified the tests to avoid testing against stored values and to test some more fundamental properties of the outputs instead. This should hopefully avoid all of the testing issues we've had so far.

Thanks,
Andrew
