### 2019-02-09 #1193 @yochannah
Okay, looking good! @ja-thomas and @schalkdaniel thanks for reviewing. 😄

@schalkdaniel - You still have a couple of checkboxes unchecked. When you get a moment could you let us know if your concerns are resolved, or if not what needs to be done? 👍
