### 2019-02-10 #1193 @yochannah
hmm, this is strange - these DOIs all seem to resolve correctly when I click on them.

I note there is a space as the last character for some of the DOIs which maybe we should fix, but that's only present on three so presumably can't be affecting them all!

- https://doi.org/10.1162/NECO_a_00813
- https://doi.org/10.1162/neco.2006.18.10.2283
- https://doi.org/10.1002/widm.1198

@andrewthomasjones - it might be a good idea to remove those spaces from https://github.com/andrewthomasjones/BoltzMM/blob/master/paper.bib just in case?

@arfon  / @openjournals/joss-eics - any idea what has whedon grumpy - something I missed or misunderstood maybe?
