+++ paper.bib
@@ -43,7 +43,7 @@
 	Journal = {WIREs Data Mining and Knowledge Discovery},
 	Pages = {e1198},
 	Title = {{An introduction to MM algorithms for machine learning and statistical estimation}},
-	Url = {https://doi.org/10.1002/widm.1198 },
+	Url = {https://doi.org/10.1002/widm.1198},
 	Volume = {7},
 	Year = {2017}}
 
@@ -78,7 +78,7 @@
 	Journal = {Neural Computation},
 	Pages = {485-492},
 	Title = {{A block successive lower-bound maximization algorithm for the maximum pseudolikelihood estimation of fully visible Boltzmann machines}},
-	Url = {https://doi.org/10.1162/NECO_a_00813 },
+	Url = {https://doi.org/10.1162/NECO_a_00813},
 	Volume = {28},
 	Year = {2016}}
 
@@ -135,7 +135,7 @@
 	Pages = {2283-2292},
 	Timestamp = {2013.12.03},
 	Title = {{Consistency of pseudolikelihood estimation of fully visible Boltzmann machines}},
-	Url = {https://doi.org/10.1162/neco.2006.18.10.2283 },
+	Url = {https://doi.org/10.1162/neco.2006.18.10.2283},
 	Volume = {18},
 	Year = {2006}}
 
