### 2019-02-10 #1193 @hiendn
Hi @yochannah, spaces have now been removed. I also clicked on all of the links and they all seem to correctly direct to the papers. So 🤷‍♂️? Thank you for you editorial work! 🙇
