### 2019-02-11 #1193 @arfon
@yochannah - I've tried fixing the whitespace issue in https://github.com/openjournals/whedon-api/commit/c1d4e5e5c0bdb343207bf19e2dee2e2e8bd7304b.
