### 2019-02-11 #1193 @arfon
More generally though, I think the issue is these should be `doi` fields, not `urls`, i.e. https://github.com/andrewthomasjones/BoltzMM/blob/master/paper.bib#L28 should be:

```
doi = {https://doi.org/10.1007/978-1-4614-6868-4}
```
