+++ paper.bib
@@ -13,7 +13,7 @@
 	Date-Added = {2019-01-12 17:14:15 +1100},
 	Date-Modified = {2019-01-12 17:46:16 +1100},
 	Title = {{deepnet: deep learning toolkit in R}},
-	Url = {https://CRAN.R-project.org/package=deepnet},
+	doi = {https://CRAN.R-project.org/package=deepnet},
 	Year = {2014}}
 
 @book{Eddelbuettel2013,
@@ -25,7 +25,7 @@
 	Publisher = {Springer},
 	Timestamp = {2015.07.01},
 	Title = {{Seamless R and C++ Integration with Rcpp}},
-	Url = {https://doi.org/10.1007/978-1-4614-6868-4},
+	doi = {https://doi.org/10.1007/978-1-4614-6868-4},
 	Year = {2013}}
 
 @manual{R-Core-Team:2018aa,
@@ -43,7 +43,7 @@
 	Journal = {WIREs Data Mining and Knowledge Discovery},
 	Pages = {e1198},
 	Title = {{An introduction to MM algorithms for machine learning and statistical estimation}},
-	Url = {https://doi.org/10.1002/widm.1198},
+	doi = {https://doi.org/10.1002/widm.1198},
 	Volume = {7},
 	Year = {2017}}
 
@@ -56,7 +56,7 @@
 	Pages = {30-37},
 	Timestamp = {2014.01.17},
 	Title = {{A tutorial on MM algorithms}},
-	Url = {https://doi.org/10.1198/0003130042836},
+	doi = {https://doi.org/10.1198/0003130042836},
 	Volume = {58},
 	Year = {2004}}
 
@@ -67,7 +67,7 @@
 	Journal = {IEEE Transactions on Neural Networks and Learning Systems},
 	Pages = {897-902},
 	Title = {{Asymptotic normality of the maximum pseudolikelihood estimator for fully visible Boltzmann machines}},
-	Url = {https://doi.org/10.1109/TNNLS.2015.2425898},
+	doi = {https://doi.org/10.1109/TNNLS.2015.2425898},
 	Volume = {27},
 	Year = {2016}}
 
@@ -78,7 +78,7 @@
 	Journal = {Neural Computation},
 	Pages = {485-492},
 	Title = {{A block successive lower-bound maximization algorithm for the maximum pseudolikelihood estimation of fully visible Boltzmann machines}},
-	Url = {https://doi.org/10.1162/NECO_a_00813},
+	doi = {https://doi.org/10.1162/NECO_a_00813},
 	Volume = {28},
 	Year = {2016}}
 
@@ -89,7 +89,7 @@
 	Journal = {Contemporary Mathematics},
 	Pages = {221-239},
 	Title = {{Composite likelihood methods}},
-	Url = {http://dx.doi.org/10.1090/conm/080/999014},
+	doi = {http://dx.doi.org/10.1090/conm/080/999014},
 	Volume = {8},
 	Year = {1988}}
 
@@ -112,7 +112,7 @@
 	Date-Modified = {2019-01-12 17:50:50 +1100},
 	Institution = {SSRN},
 	Title = {{The fully-visible Boltzmann machine and the Senate of the 45th Australian Parliament in 2016}},
-	Url = {http://dx.doi.org/10.2139/ssrn.3287903},
+	doi = {http://dx.doi.org/10.2139/ssrn.3287903},
 	Year = {2018}}
 
 @article{Cox:1972aa,
@@ -135,7 +135,7 @@
 	Pages = {2283-2292},
 	Timestamp = {2013.12.03},
 	Title = {{Consistency of pseudolikelihood estimation of fully visible Boltzmann machines}},
[…]
