+++ paper.bib
@@ -13,7 +13,7 @@
 	Date-Added = {2019-01-12 17:14:15 +1100},
 	Date-Modified = {2019-01-12 17:46:16 +1100},
 	Title = {{deepnet: deep learning toolkit in R}},
-	doi = {https://CRAN.R-project.org/package=deepnet},
+	Url = {https://CRAN.R-project.org/package=deepnet},
 	Year = {2014}}
 
 @book{Eddelbuettel2013,
