### 2019-02-11 #1193 @hiendn
Hi @arfon and @yochannah, I've changed everything that should be a DOI from URL to DOI. There are a couple of items left that do not have DOIs, so I left them as URL. I hope that this resolves the issue! 😄
