+++ paper.bib
@@ -122,7 +122,7 @@
 	Journal = {Journal of the Royal Statistical Society C},
 	Number = {113-120},
 	Title = {{The analysis of multivariate binary data}},
-	Url = {https://www.jstor.org/stable/2346482},
+	doi = {https://doi.org/10.2307/2346482},
 	Volume = {21},
 	Year = {1972}}
 
