### 2019-02-11 #1193 @hiendn
@arfon , it doesn't look like a DOI exists anywhere on the internet for that last paper, unfortunately. 😞
