### 2019-02-11 #1193 @arfon
> @arfon , it doesn't look like a DOI exists anywhere on the internet for that last paper, unfortunately.

That's OK. These are only meant to be suggestions - it's fine to have references without DOIs if there isn't one.
