### 2019-02-12 #1193 @yochannah
@hiendn @andrewthomasjones - at this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive?  It's probably worth giving the manuscript a final proofread if you feel you need to, as well. 👍
