+++ paper.md
@@ -55,7 +55,7 @@ All of the functions that have been comprehensively documented, and the document
 
 The `BoltzMM` package is programmed in native `R`, with particularly computationally intensive subroutines programmed in `C` and integrated via the `Rcpp` and `RcppArmadillo` packages of @Eddelbuettel2013. The package is therefore sufficiently suitable for analysis of moderately large dimensional data sets. On https://cran.r-project.org, the only other package that provides facilities for estimation of BMs, of any kind, is the `deepnet` package of @Rong:2014aa. Here, facilities for estimation of the so-called *restricted BM* (RBM) are included, via the functions `rbm.down`, `rbm.train`, and `rbm.up`. The RBM is an alternative simplification to the BM, which preserves the denseness properties BM but can be parameterized by fewer parameter elements. However, the estimation of RBM models is as intractable as for BM models, and due to the latent variable construction preventing the establishment of distributional results for the estimators of the RBM parameter elements, its use for statistical inference remains limited. Thus we view the functions from `deepnet` as being complimentary but not overlapping with `BoltzMM`.
 
-Users can obtain the latest build of `BoltzMM` on GitHub (https://github.com/andrewthomasjones/BoltzMM). The latest stable build can be obtained from CRAN (https://CRAN.R-project.org/package=BoltzMM), and an archival build can be obtained from Zenodo (http://doi.org/10.5281/zenodo.2538256). Algorithm derivations and theoretical results regarding the implented functions can be found in @Nguyen2016 and @NguyenWood2016. An example application of the `BoltzMM` package for the analysis of real data is comprehensively described in @Bagnall:2018aa.   Thorough descriptions of the package functions appear in the manual, which can be accessed at https://cran.r-project.org/web/packages/BoltzMM/BoltzMM.pdf. Bug reports and other feedback can be directed to the GitHub issues page (https://github.com/andrewthomasjones/BoltzMM/issues).
+Users can obtain the latest build of `BoltzMM` on GitHub (https://github.com/andrewthomasjones/BoltzMM). The latest stable build can be obtained from CRAN (https://CRAN.R-project.org/package=BoltzMM), and an archival build can be obtained from Zenodo (http://doi.org/10.5281/zenodo.2563411). Algorithm derivations and theoretical results regarding the implented functions can be found in @Nguyen2016 and @NguyenWood2016. An example application of the `BoltzMM` package for the analysis of real data is comprehensively described in @Bagnall:2018aa.   Thorough descriptions of the package functions appear in the manual, which can be accessed at https://cran.r-project.org/web/packages/BoltzMM/BoltzMM.pdf. Bug reports and other feedback can be directed to the GitHub issues page (https://github.com/andrewthomasjones/BoltzMM/issues).
 
 # Acknowledgements
 Jessica Bagnall is funded by a Research Training Program Stipend scholarship from La Trobe University. Hien Nguyen is funded under Australian Research Council (ARC) grant numbers DE170101134 and DP180101192.
