### 2019-02-13 #1193 @labarba
👋  @hiendn — Could you edit the title in Zenodo so it matches the title of the JOSS paper? Thanks!
