### 2019-02-14 24de71e
CRAN elease documentation update
