### 2019-02-14 #1193 @labarba
@hiendn — I'm going to have a few more editorial fixes to request.

¶1
that is able densely represent >> is able _to_
when d is large and sample size n >> what about the sample size? (missing a verb and parallel construction)

In general, I find the whole first page really jargon-heavy and I wonder if you would consider editing this to be more readable. E.g., what is "latent variable construction"? Do you mean the same by "complex form" and "elaborate form" And what _do_ you mean by that? Jus that it's hard to implement? Or is it really "intractable"? Because then you talk about a "tractable simplification" ... so many big words. Are they all necessary for the JOSS audience to plough through, to finally learn what the software does on page 2?

Many of your DOIs are not resolving because you have a full URL instead of the DOI in the BibTeX field, so the link appears broken. Please check.
