### 2019-02-14 #1193 @labarba
Page 2:
All of the functions that have been comprehensively documented >> delete "that"
Furthermore, the analysis of data arising from… >> Unclear how this sentence is connected with the previous and the need for "Furthermore"
complimentary >> complementary (meaning: flattering versus supportive)

It's not necessary to include the GitHub repo address or the archive address in the text, because they will appear linked in the page-1 margin of the formatted paper, and will also be linked on the paper web page. Please remove.

Overall:
I suggest a pass of ruthless editing for clarity and conciseness. For example, how does one know the difference between "comprehensively documented" and "documented," between a report that "describes" versus "clearly describes"? What is "particularly computationally intensive" versus just "computationally intensive"? When is a package "sufficiently suitable" versus "suitable"?

What I'm trying to say, as an editor, is that the paper is too long and a slog to read, and I hope you can give it a bit of a clean.

Also, note the [Author Guideline](https://joss.readthedocs.io/en/latest/submitting.html) that says to "include in the paper some sentences that would explain the software functionality and domain of use to a non-specialist reader."
