### 2019-02-15 bff428b
New version of paper
