+++ paper.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% http://bibdesk.sourceforge.net/
 
-%% Created for Hien Nguyen at 2019-01-12 17:50:59 +1100 
+%% Created for Hien Nguyen at 2019-02-15 14:18:39 +1100 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -20,12 +20,12 @@
 	Address = {New York},
 	Author = {D Eddelbuettel},
 	Date-Added = {2019-01-12 17:12:58 +1100},
-	Date-Modified = {2019-01-12 17:49:53 +1100},
+	Date-Modified = {2019-02-15 10:44:31 +1100},
+	Doi = {10.1007/978-1-4614-6868-4},
 	Owner = {uqhngu10},
 	Publisher = {Springer},
 	Timestamp = {2015.07.01},
 	Title = {{Seamless R and C++ Integration with Rcpp}},
-	doi = {https://doi.org/10.1007/978-1-4614-6868-4},
 	Year = {2013}}
 
 @manual{R-Core-Team:2018aa,
@@ -39,126 +39,113 @@
 @article{Nguyen:2017aa,
 	Author = {H D Nguyen},
 	Date-Added = {2019-01-12 16:12:33 +1100},
-	Date-Modified = {2019-01-12 17:43:17 +1100},
+	Date-Modified = {2019-02-15 10:45:05 +1100},
+	Doi = {10.1002/widm.1198},
 	Journal = {WIREs Data Mining and Knowledge Discovery},
 	Pages = {e1198},
 	Title = {{An introduction to MM algorithms for machine learning and statistical estimation}},
-	doi = {https://doi.org/10.1002/widm.1198},
 	Volume = {7},
 	Year = {2017}}
 
 @article{Hunter2004,
 	Author = {D R Hunter and K Lange},
 	Date-Added = {2019-01-12 16:12:07 +1100},
-	Date-Modified = {2019-01-12 17:42:52 +1100},
+	Date-Modified = {2019-02-15 10:44:38 +1100},
+	Doi = {10.1198/0003130042836},
 	Journal = {The American Statistician},
 	Owner = {hiennguyen},
 	Pages = {30-37},
 	Timestamp = {2014.01.17},
 	Title = {{A tutorial on MM algorithms}},
-	doi = {https://doi.org/10.1198/0003130042836},
 	Volume = {58},
 	Year = {2004}}
 
 @article{Nguyen2016,
 	Author = {H D Nguyen and I A Wood},
 	Date-Added = {2019-01-12 16:08:55 +1100},
-	Date-Modified = {2019-01-12 17:44:21 +1100},
+	Date-Modified = {2019-02-15 10:45:11 +1100},
+	Doi = {10.1109/TNNLS.2015.2425898},
 	Journal = {IEEE Transactions on Neural Networks and Learning Systems},
 	Pages = {897-902},
 	Title = {{Asymptotic normality of the maximum pseudolikelihood estimator for fully visible Boltzmann machines}},
-	doi = {https://doi.org/10.1109/TNNLS.2015.2425898},
 	Volume = {27},
 	Year = {2016}}
 
 @article{NguyenWood2016,
 	Author = {H D Nguyen and I A Wood},
 	Date-Added = {2019-01-12 16:08:49 +1100},
-	Date-Modified = {2019-01-12 17:41:53 +1100},
+	Date-Modified = {2019-02-15 10:45:18 +1100},
+	Doi = {10.1162/NECO_a_00813},
 	Journal = {Neural Computation},
 	Pages = {485-492},
[…]
