### 2019-02-15 #1193 @hiendn
Hi @labarba, thank you for the very comprehensive technical editing of the paper! 🙇 Your careful reading and many suggestions have been incredibly helpful in my reorganisation of the paper to be more suitable for the lay audience of the journal. Please find my responses to your comments, below. I hope that I have addressed all of your concerns . 😄

**@hiendn — I'm going to have a few more editorial fixes to request.

¶1
that is able densely represent >> is able _to_**

---In light of your comments below, regarding jargon, I have chosen to replace the word densely represent, with 'closely approximate', which I feel can be interpreted at face value by a lay reader, without the need for a dive into approximation theoretic nomenclature. The grammar of the sentence has also been fixed in order to comply with the new phraseology.

**when d is large and sample size n >> what about the sample size? (missing a verb and parallel construction)**

---Thanks for pointing out the typo, I have now change it to 'when both dimension $d$ and sample size $n$ are large.'

**In general, I find the whole first page really jargon-heavy and I wonder if you would consider editing this to be more readable.**

---Thank you for this comment. I suppose that I am new to writing more lay focused papers and so a lot of my technical writing habits come through. In what follows, I endeavour to choose better words that are meaningful to a lay audience, but do not detract from the technical meaning, either.

**E.g., what is "latent variable construction"?**

---'Latent variable construction' is now replaced by 'requirement for hidden (or unobserved) variables'.

**Do you mean the same by "complex form" and "elaborate form" And what _do_ you mean by that? Jus that it's hard to implement? Or is it really "intractable"? Because then you talk about a "tractable simplification" ... so many big words.**

---You are correct in inferring that all of these phrases are just my attempt to not be repetitively writing 'intractable' over and over again. Although, with your comments in mind, I have now combed the document for unnecessary instances of words that could be replaced by things with more meaning to the reader.

**Are they all necessary for the JOSS audience to plough through, to finally learn what the software does on page 2?**

---I have rearranged the paragraphs of the article to be more direct in presenting the core value of the software to the reader, before going on a tangent with what the model that underlies the software is, and its history.

**Many of your DOIs are not resolving because you have a full URL instead of the DOI in the BibTeX field, so the link appears broken. Please check.
[EDIT] To clarify, the links come out as `https://doi.org/https://doi.org/10.1016/S0364-0213(85)80012-4` which don't resolve.**

---The http... bits of the string, up to the DOIs have now been removed from each of the references. Thank you for this correction. I didn't know how the system was handling the DOIs and thought that all of the OKs from the reference checks meant that I must have been doing the right thing.

**Page 2:
All of the functions that have been comprehensively documented >> delete "that"**

---'that' has been deleted.

**Furthermore, the analysis of data arising from… >> Unclear how this sentence is connected with the previous and the need for "Furthermore"**

---'Furthermore' has been removed.

**complimentary >> complementary (meaning: flattering versus supportive)**

---I get these two wrong all of the time. Thank you for the correction!

**It's not necessary to include the GitHub repo address or the archive address in the text, because they will appear linked in the page-1 margin of the formatted paper, and will also be linked on the paper web page. Please remove.**

---Github repo and archive addresses have been removed.

**Overall:
I suggest a pass of ruthless editing for clarity and conciseness. For example, how does one know the difference between "comprehensively documented" and "documented," between a report that "describes" versus "clearly describes"? What is "particularly computationally intensive" versus just "computationally intensive"? When is a package "sufficiently suitable" versus "suitable"?**

---You've caught me out on one of my quirks of being overly generous with adjectives and adverbs usage. The document has been combed for the over abundance of such phraseology.

**What I'm trying to say, as an editor, is that the paper is too long and a slog to read, and I hope you can give it a bit of a clean.**

---Point taken. I have rearranged the document so that the interesting software stuff is at the start and the more technical stuff, which I would still like to include for completeness, appears afterwards.

**Also, note the [Author Guideline](https://joss.readthedocs.io/en/latest/submitting.html) that says to "include in the paper some sentences that would explain the software functionality and domain of use to a non-specialist reader."**

---I have included the following sentence for non-specialist readers: "Simply put, the FVBM is a model that allows for the characterization of interactions between multiple binary random variables, simultaneously."
