### 2019-02-15 #1193 @labarba
Thanks, @hiendn.

¶1
The FVBM was  first in Hyvarinen (2006) >> missing verb?
