### 2019-02-15 #1193 @hiendn
@labarba haha thanks for spotting the typo; 'considered' was the missing verb. Now included. Its amazing what you end up mentally filling in when you proof read your own writing.
